\name{SegAnnot-package}
\alias{SegAnnot-package}
\alias{SegAnnot}
\docType{package}
\title{Annotation-aware segmentation via dynamic programing}
\description{The SegAnnot package provides a function
'SegAnnot()' that calculates the L2 optimal segmentation
for a given set of 1-annotated regions.}
\details{
\tabular{ll}{Package: \tab SegAnnot\cr
Version: \tab 1.1\cr
Date: \tab 2013-01-25\cr
Title: \tab Annotation-aware segmentation via dynamic programing\cr
License: \tab Unlimited\cr
Author: \tab Guillem Rigaill, Toby Dylan Hocking\cr
Maintainer: \tab Guillem Rigaill <g.rigaill@nki.nl>\cr
Depends: \tab R (>= 2.10)\cr
Suggests: \tab proto, ggplot2, cghseg, tikzDevice\cr}
}
\author{Guillem Rigaill, Toby Dylan Hocking}

\keyword{ package }


