\name{smoothedSignal}
\alias{smoothedSignal}
\title{smoothedSignal}
\description{Calculate the vector of smoothed observations.}
\usage{
smoothedSignal(x,
saut,
method = mean)
}
\arguments{
  \item{x}{
}
  \item{saut}{
}
  \item{method}{
}
}



\author{Guillem Rigaill, Toby Dylan Hocking}





